Added lombok to all remaining models.
