+++ pom.xml
@@ -81,7 +81,6 @@
 <dependency>
             <groupId>org.projectlombok</groupId>
             <artifactId>lombok</artifactId>
-            <optional>true</optional>
         </dependency>
 		
 		<dependency>
