+++ src/main/java/com/revature/controllers/AdminController.java
@@ -17,7 +17,7 @@ import org.springframework.web.bind.annotation.RequestBody;
 import org.springframework.web.bind.annotation.RequestMapping;
 import org.springframework.web.bind.annotation.RestController;
 
-import com.revature.model.Admin;
+import com.revature.models.Admin;
 import com.revature.services.AdminService;
 
 import io.swagger.annotations.Api;
