+++ src/main/java/com/revature/controllers/LoginController.java
@@ -26,7 +26,7 @@ import org.springframework.web.bind.annotation.RequestMethod;
 import org.springframework.web.bind.annotation.RequestParam;
 import org.springframework.web.bind.annotation.RestController;
 
-import com.revature.model.User;
+import com.revature.models.User;
 import com.revature.services.DistanceService;
 import com.revature.services.UserService;
