+++ src/main/java/com/revature/models/Admin.java
@@ -1,4 +1,4 @@
-package com.revature.model;
+package com.revature.models;
 
 import java.io.Serializable;
