+++ src/main/java/com/revature/models/package-info.java
@@ -2,4 +2,4 @@
  * Package that contains all beans.
  */
 
-package com.revature.model;
+package com.revature.models;
