+++ src/main/java/com/revature/repositories/BatchRepository.java
@@ -6,7 +6,7 @@ import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.data.jpa.repository.Query;
 import org.springframework.stereotype.Repository;
 
-import com.revature.model.Batch;
+import com.revature.models.Batch;
 
 /**
  * BatchRepository which extends the JpaRepository.
