+++ src/main/java/com/revature/repositories/CarRepository.java
@@ -4,7 +4,7 @@ import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.data.jpa.repository.Query;
 import org.springframework.stereotype.Repository;
 
-import com.revature.model.Car;
+import com.revature.models.Car;
 
 /**
  * CarRepository which extends the JpaRepository.
