+++ src/main/java/com/revature/repositories/UserRepository.java
@@ -6,7 +6,7 @@ import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.data.jpa.repository.Query;
 import org.springframework.stereotype.Repository;
 
-import com.revature.model.User;
+import com.revature.models.User;
 
 /**
  * UserRepository which extends the JpaRepository.
