+++ src/main/java/com/revature/services/BatchService.java
@@ -2,7 +2,7 @@ package com.revature.services;
 
 import java.util.List;
 
-import com.revature.model.Batch;
+import com.revature.models.Batch;
 
 public interface BatchService {
