+++ src/main/java/com/revature/services/CarService.java
@@ -2,7 +2,7 @@ package com.revature.services;
 
 import java.util.List;
 
-import com.revature.model.Car;
+import com.revature.models.Car;
 
 public interface CarService {
