+++ src/main/java/com/revature/services/UserService.java
@@ -2,7 +2,7 @@ package com.revature.services;
 
 import java.util.List;
 
-import com.revature.model.User;
+import com.revature.models.User;
 
 public interface UserService {
