+++ src/main/java/com/revature/services/impl/BatchServiceImpl.java
@@ -5,7 +5,7 @@ import java.util.List;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Service;
 
-import com.revature.model.Batch;
+import com.revature.models.Batch;
 import com.revature.repositories.BatchRepository;
 import com.revature.services.BatchService;
