+++ src/main/java/com/revature/services/impl/UserServiceImpl.java
@@ -5,7 +5,7 @@ import java.util.List;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Service;
 
-import com.revature.model.User;
+import com.revature.models.User;
 import com.revature.repositories.UserRepository;
 import com.revature.services.UserService;
