+++ src/test/java/com/revature/controllers/BatchControllerTest.java
@@ -22,7 +22,7 @@ import org.springframework.test.context.junit4.SpringRunner;
 import org.springframework.test.web.servlet.MockMvc;
 
 import com.fasterxml.jackson.databind.ObjectMapper;
-import com.revature.model.Batch;
+import com.revature.models.Batch;
 import com.revature.services.BatchService;
 
 @RunWith(SpringRunner.class)
