+++ src/test/java/com/revature/controllers/CarControllerTest.java
@@ -22,8 +22,8 @@ import org.springframework.test.context.junit4.SpringRunner;
 import org.springframework.test.web.servlet.MockMvc;
 
 import com.fasterxml.jackson.databind.ObjectMapper;
-import com.revature.model.Car;
-import com.revature.model.User;
+import com.revature.models.Car;
+import com.revature.models.User;
 import com.revature.services.CarService;
 
 @RunWith(SpringRunner.class)
